completed

+++ src/modules/url/repositories/url.repository.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { UrlDocument, UrlEntity } from '../../../core/entities/url.entity';
 import { IUrlRepository } from '../interfaces/url-repository.interface';
-import { Model } from 'mongoose'
+import { Model, Types } from 'mongoose'
 import { InjectModel } from '@nestjs/mongoose';
 
 
@@ -22,18 +22,19 @@ export class UrlRepository implements IUrlRepository {
   }
 
   async findAllByUserId(userId: string): Promise<UrlDocument[]> {
-    return this.urlModel.find({ userId }).exec();
+    // return this.urlModel.find().exec();
+    return this.urlModel.find({ userId: new Types.ObjectId(userId) }).exec();
   }
 
   async incrementVisitCount(id: string): Promise<void> {
     await this.urlModel.updateOne(
-      { _id: id },
+      { _id: new Types.ObjectId(id) },
       { $inc: { visitCount: 1 } },
     ).exec();
   }
 
   async delete(id: string, userId: string): Promise<boolean> {
-    const result = await this.urlModel.deleteOne({ _id: id, userId }).exec();
+    const result = await this.urlModel.deleteOne({ _id: id, userId : new Types.ObjectId(userId) }).exec();
     return result.deletedCount > 0;
   }
 }
